\name{gofBridge-package}
\alias{gofBridge-package}
\alias{gofBridge}
\docType{package}
\title{Tests d'adéquation aux processus de dégradation gamma et Wiener homogènes}
\description{
Tests d'adéquation au processus gamma homogène et au processus de Wiener
homogène observés à des instants réguliers ou irréguliers. La fonction
principale est \code{\link{Bridge}} ; les fonctions \code{\link{sim_HGP}},
\code{\link{sim_HWP}} et suivantes simulent des trajectoires sous H0 et sous
des alternatives.
}
\author{TITO A. Rytrace \email{rytracetito5@gmail.com}}
\keyword{package}
